requirements.txt
torch==2.1.0
transformers==4.36.2
diffusers==0.25.0
accelerate
scipy
Pillow